move file to common directory
